+++ src/github/github2.test.ts
@@ -1,28 +1,10 @@
 import { Context } from "@actions/github/lib/context";
 import { replayOctokit } from "../replay";
-import { type Octokit, getPRLabels, getPRsLabels, getWorkflowRunJobs } from "./github";
+import { type Octokit, getPRLabels, getPRsLabels } from "./github";
 
 const token = process.env["GH_TOKEN"] ?? "";
 const owner = "corentinmusard";
 const repo = "otel-cicd-action";
-
-describe("getWorkflowRunJobs", () => {
-  let octokit: Octokit;
-
-  beforeAll(async () => {
-    process.env["GITHUB_REPOSITORY"] = "biomejs/biome";
-    octokit = await replayOctokit("getWorkflowRunJobs", token);
-  });
-
-  it("should return the workflow run jobs", async () => {
-    const context = new Context();
-    context.runId = 2; // different than 12541749172
-
-    const workflowRunJobs = await getWorkflowRunJobs(context, octokit, 12541749172);
-    expect(workflowRunJobs.jobs).toHaveLength(8);
-    expect(workflowRunJobs.workflowRun.id).toBe(12541749172);
-  }, 10000);
-});
 
 describe("getPRLabels", () => {
   let octokit: Octokit;