+++ src/runner.ts
@@ -1,6 +1,6 @@
 import * as core from "@actions/core";
 import { context, getOctokit } from "@actions/github";
-import { getPRsLabels, getWorkflowRunJobs } from "./github/github";
+import { getPRsLabels, getWorkflowRun, listJobsForWorkflowRun, listWorkflowRunArtifacts } from "./github/github";
 import { traceWorkflowRun } from "./tracing/job";
 import { type Attributes, createTracerProvider } from "./tracing/trace";
 
@@ -12,31 +12,39 @@ async function run() {
   const ghToken = core.getInput("githubToken") || process.env["GITHUB_TOKEN"] || "";
   const octokit = getOctokit(ghToken);
 
-  core.info(`Get workflow run jobs for ${runId}`);
-  const workflowRunJobs = await getWorkflowRunJobs(context, octokit, runId);
+  core.info(`Get workflow run for ${runId}`);
+  const workflowRun = await getWorkflowRun(context, octokit, runId);
+
+  core.info("Get artifacts");
+  const artifacts = await listWorkflowRunArtifacts(context, octokit, runId);
+
+  core.info("Get jobs");
+  const jobs = await listJobsForWorkflowRun(context, octokit, runId);
 
   core.info("Get PRs labels");
-  const prNumbers = workflowRunJobs.workflowRun.pull_requests?.map((pr) => pr.number) ?? [];
+  const prNumbers = workflowRun.pull_requests?.map((pr) => pr.number) ?? [];
   const prLabels = await getPRsLabels(context, octokit, prNumbers);
 
   core.info(`Create tracer provider for ${otlpEndpoint}`);
   const attributes: Attributes = {
-    serviceName: otelServiceName || workflowRunJobs.workflowRun.name || `${workflowRunJobs.workflowRun.workflow_id}`,
-    serviceVersion: workflowRunJobs.workflowRun.head_sha,
+    serviceName: otelServiceName || workflowRun.name || `${workflowRun.workflow_id}`,
+    serviceVersion: workflowRun.head_sha,
     serviceInstanceId: [
-      workflowRunJobs.workflowRun.repository.full_name,
-      `${workflowRunJobs.workflowRun.workflow_id}`,
-      `${workflowRunJobs.workflowRun.id}`,
-      `${workflowRunJobs.workflowRun.run_attempt ?? 1}`,
+      workflowRun.repository.full_name,
+      `${workflowRun.workflow_id}`,
+      `${workflowRun.id}`,
+      `${workflowRun.run_attempt ?? 1}`,
     ].join("/"),
-    serviceNamespace: workflowRunJobs.workflowRun.repository.full_name,
+    serviceNamespace: workflowRun.repository.full_name,
   };
   const provider = createTracerProvider(otlpEndpoint, otlpHeaders, attributes);
 
   try {
     core.info(`Trace workflow run for ${runId} and export to ${otlpEndpoint}`);
-    const traceId = await traceWorkflowRun(workflowRunJobs, prLabels);
+    const traceId = await traceWorkflowRun(workflowRun, jobs, artifacts, prLabels);
+
     core.setOutput("traceId", traceId);
+    core.debug(`traceId: ${traceId}`);
 
     core.info("Flush and shutdown tracer provider");
     await provider.forceFlush();