+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.13.2] - 2025-01-04
+
+### Fixed
+
+- Fix artifacts parsing for `@opentelemetry/exporter-trace-otlp-http` version `<0.29.0`
+
+### Changed
+
+- Refactor artifacts handling
+
 ## [1.13.1] - 2024-12-31
 
 ### Fixed
@@ -113,7 +123,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `https` endpoints (proto over http).
 - Update to node 20.x
 
-[unreleased]: https://github.com/corentinmusard/otel-cicd-action/compare/v1.13.1...HEAD
+[unreleased]: https://github.com/corentinmusard/otel-cicd-action/compare/v1.13.2...HEAD
+[1.13.2]: https://github.com/corentinmusard/otel-cicd-action/compare/v1.13.1...v1.13.2
 [1.13.1]: https://github.com/corentinmusard/otel-cicd-action/compare/v1.13.0...v1.13.1
 [1.13.0]: https://github.com/corentinmusard/otel-cicd-action/compare/v1.12.1...v1.13.0
 [1.12.1]: https://github.com/corentinmusard/otel-cicd-action/compare/v1.12.0...v1.12.1